* Added fancy (ie. most people would say bad idea) WrapWithGo function.
Call it on a function that accepts one arg; it attaches a "Go" setter onto the function, which you can call:
    MyFunc.Go = argument1Value;
which is equivalent to:
    MyFunc(argument1Value);

Used for quick tests and such, eg. for calling something on a multiline value like a callback, without having to add a parenthesis at the end to mark the end of the function call.