* Added defensive call to ClearIntervalOrTimeout() just before each of the two Timer class' code-lines that starts a new interval/timeout.
Today (2024-09-21), a case seems to have occurred where, two instances of an interval were running for a timer at the same time. (initialDelayOverride usage was involved, if that matters) So I added these checks to avoid this happening again. (I don't know what exact call-chain led to this situation, but I'm guessing it was some line through the Timer class' code I hadn't considered earlier; we'll see if my changes fix it / prevent an eventual recurrence)